Backport r1173689
Added a "steps" parameter to the animations, that limits the maximum number of different 
values a given animation can take. This should effectively reduce the size of the caches by 
a factor 10 to 20, improve performances, without any noticeable differences to the eye, and 
possibly fix some of the performance issues experienced by some users with NVidia graphic 
cards. 

The number of steps is configurable (as an hidden option) and is still to be optimized. 

Current value is set to 10. (was 256 in the past)


svn path=/trunk/playground/artwork/oxygen-transparent/; revision=1174508